fix: remove nested conftest.py
